fix(pre-s6-cleanup): BigInteger migration, is_active check, pagination, tech-debt resolved (TD-016/011/003/009)

+++ app/models/entry.py
@@ -12,7 +12,7 @@ import enum
 import uuid
 from typing import TYPE_CHECKING
 
-from sqlalchemy import BigInteger, CheckConstraint, Enum, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, CheckConstraint, Enum, ForeignKey, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
@@ -52,7 +52,7 @@ class Entry(Base):
         Enum(Direction, name="direction"),
         nullable=False,
     )
-    amount: Mapped[int] = mapped_column(Integer, nullable=False)
+    amount: Mapped[int] = mapped_column(BigInteger, nullable=False)
     currency: Mapped[str] = mapped_column(
         String(3),
         nullable=False,  # ISO 4217 code e.g. "EUR", "USD", "JPY"
